+++ phoenix-knowledge-provider/src/main/java/com/ginkgocap/ywxt/knowledge/service/impl/ColumnVisibleServiceImpl.java
@@ -141,7 +141,7 @@ public class ColumnVisibleServiceImpl implements ColumnVisibleService {
             cv.setUtime(new Date());
             cv.setSortId(c.getColumnLevelPath());
             cv.setColumnName(cname);
-            cv.setState((byte) 0);
+            cv.setState((short) 0);
             columnVisibleMapper.insert(cv);
         }
     }
@@ -160,7 +160,7 @@ public class ColumnVisibleServiceImpl implements ColumnVisibleService {
         cv.setUtime(new Date());
         cv.setColumnName(cname);
         cv.setSortId(c.getColumnLevelPath());
-        cv.setState((byte) 0);
+        cv.setState((short) 0);
         columnVisibleMapper.insert(cv);
     }
